+++ dataforge-context/src/commonMain/kotlin/space/kscience/dataforge/context/Global.kt
@@ -3,7 +3,7 @@ package space.kscience.dataforge.context
 import kotlinx.coroutines.CoroutineName
 import kotlinx.coroutines.Job
 import space.kscience.dataforge.meta.Meta
-import space.kscience.dataforge.names.asName
+import space.kscience.dataforge.names.Name
 import space.kscience.dataforge.names.parseAsName
 import kotlin.coroutines.CoroutineContext
 import kotlin.native.concurrent.ThreadLocal
@@ -14,7 +14,7 @@ internal expect fun getGlobalLoggerFactory(): PluginFactory<out LogManager>
  * A global root context. Closing [Global] terminates the framework.
  */
 @ThreadLocal
-private object GlobalContext : Context("GLOBAL".asName(), null, emptySet(), Meta.EMPTY) {
+private object GlobalContext : Context(Name.of("GLOBAL"), null, emptySet(), Meta.EMPTY) {
     override val coroutineContext: CoroutineContext = Job() + CoroutineName("GlobalContext")
 }