+++ dataforge-meta/src/commonTest/kotlin/space/kscience/dataforge/meta/MutableMetaViewTest.kt
@@ -1,6 +1,6 @@
 package space.kscience.dataforge.meta
 
-import space.kscience.dataforge.names.asName
+import space.kscience.dataforge.names.Name
 import kotlin.test.Test
 import kotlin.test.assertEquals
 import kotlin.test.assertTrue
@@ -9,7 +9,7 @@ class MutableMetaViewTest {
     @Test
     fun metaView() {
         val meta = MutableMeta()
-        val view = meta.view("a".asName())
+        val view = meta.view(Name.of("a"))
 
         view["b"] = Meta.EMPTY